backend/requirements.txt
fastapi
uvicorn
websockets
faster-whisper
numpy
pydub
sentence-transformers
faiss-cpu
groq
httpx
